Add JSON handlebars helper
Adds a helper to dump a variable as JSON. This is intended as a debugging
aid and likely not for use in an actual contract. The variable is NOT
wrapped with a `div` and is formatted as a code block.

Renames CONTRIBUTING.MD to .md to fix the broken link in the README
and to be consistent with other EOSIO projects.

Updates the `front-matter` lib to the latest version.