Add `if_has_value` helper
The `if_has_value` block helper will render a block if the given variable
is defined AND non-null. In the case of undefined or null value, nothing
is rendered, or the `{{else}}` block renders if specified.

Also moved the `as_json` helper to the 0.2 version as that is a foreward
compatible addition, but shouldn't be added to the old version processors
per the specification.

Updated generated docs.